requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.6
django-etc==1.3.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
pkg-resources==0.0.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
